Setting thought model up
